requirements.txt
django~=3.0.0
pillow
django-extension
django-bootstrap4
django-debug-toolbar